Split out cf-specific code from api.types.ts
