+++ src/frontend/packages/core/src/core/user-favorite-helpers.ts
@@ -1,13 +1,14 @@
-import { CfAPIResource } from '../../../store/src/types/api.types';
 import { EndpointModel } from '../../../store/src/types/endpoint.types';
 import { IFavoriteMetadata, UserFavorite, UserFavoriteEndpoint } from '../../../store/src/types/user-favorites.types';
 import { IEntityMetadata } from './entity-catalogue/entity-catalogue.types';
 import { FavoritesConfigMapper } from '../shared/components/favorites-meta-card/favorite-config-mapper';
+import { CfAPIResource } from '../../../cloud-foundry/src/store/types/cf-api.types';
 
 export function isEndpointTypeFavorite(favorite: UserFavorite<IFavoriteMetadata>) {
   return !favorite.entityId;
 }
 
+// TODO: This filew has cf specific code in it - this needs to be moved out
 export function getFavoriteFromCfEntity<T extends IEntityMetadata = IEntityMetadata>(
   entity,
   entityKey: string,
