+++ src/frontend/packages/store/src/types/api.types.ts
@@ -12,13 +12,6 @@ export interface APIResource<T = any> {
   entity: T;
 }
 
-export interface CfAPIResource<T = any> extends APIResource {
-  entity: {
-    [entityKey: string]: any,
-    cfGuid: string
-  };
-}
-
 export function instanceOfAPIResource(object: any): boolean {
   return object && typeof object === 'object' && 'metadata' in object && 'entity' in object;
 }
@@ -37,24 +30,7 @@ export interface NormalizedResponse<T = any> {
 
 export type ActionMergeFunction = (oldEntities: BaseEntityValues, newEntities: NormalizedResponseEntities)
   => NormalizedResponseEntities;
+
 export interface NormalizedResponseEntities {
   [key: string]: string;
-}
-
-export function createEmptyCfResponse<T = any>(): CFResponse<T> {
-  return {
-    total_results: 0,
-    total_pages: 0,
-    prev_url: '',
-    next_url: '',
-    resources: new Array<APIResource<T>>()
-  };
-}
-
-export interface CFResponse<T = any> {
-  total_results: number;
-  total_pages: number;
-  prev_url: string;
-  next_url: string;
-  resources: APIResource<T>[];
 }
