+++ src/test-e2e/marketplace/services-helper-e2e.ts
@@ -1,7 +1,8 @@
 import { browser, promise, protractor } from 'protractor';
 
+import { CFResponse, createEmptyCfResponse } from '../../frontend/packages/cloud-foundry/src/store/types/cf-api.types';
 import { IServiceInstance } from '../../frontend/packages/core/src/core/cf-api-svc.types';
-import { APIResource, CFResponse, createEmptyCfResponse } from '../../frontend/packages/store/src/types/api.types';
+import { APIResource } from '../../frontend/packages/store/src/types/api.types';
 import { e2e, E2ESetup } from '../e2e';
 import { CFHelpers } from '../helpers/cf-helpers';
 import { CFRequestHelpers } from '../helpers/cf-request-helpers';
